Inline event-options
